Multi-interface services
Sometimes is it necessary to support multiple interfaces on a single
D-Bus object. This commit extends the `Skeleton` class template to allow
multiple interfaces.

It can similarly be useful to simply rely on the built-in support for
various FreeDesktop interfaces, so the `Skeleton<>` class template
should allow you have _no_ interfaces.