Using properly sized types when ever possible
Based on the dbus documentation [0] we get concrete requirements of how large specific types have to be.
Example: t -> Unsigned 64-bit integer
Before we were using `unsigned long long` which by definition is **at least** 64 bit.
So on some systems this type could be 72 bits if the implementation sees need for it [1].

[0]: https://dbus.freedesktop.org/doc/dbus-specification.html

[1]: https://stackoverflow.com/questions/5836329/how-many-bytes-is-unsigned-long-long